stack: fix time_t printf compilation warning
Compiling on a Raspberry Pi procuced this warning:

    ../stack/source/6lowpan/ws/ws_pae_controller.c: In function ‘ws_pae_controller_frame_counter_store’:
    ../stack/source/6lowpan/ws/ws_pae_controller.c:1628:29: warning: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument 3 has type ‘time_t’ {aka ‘long int’} [-Wformat=]
     1628 |         fprintf(info->file, "# stored time: %" PRIu64 "\n", time_current(CLOCK_REALTIME));
          |                             ^~~~~~~~~~~~~~~~~~              ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
          |                                                             |
          |                                                             time_t {aka long int}
    In file included from ../stack/source/6lowpan/ws/ws_pae_controller.c:23:
    /usr/include/inttypes.h:105:34: note: format string is defined here
      105 | # define PRIu64  __PRI64_PREFIX "u"

For platform dependent types such as time_t, common practice is to cast
to their largest possible type when printing.

+++ stack/source/6lowpan/ws/ws_pae_controller.c
@@ -1625,7 +1625,7 @@ static void ws_pae_controller_frame_counter_store(pae_controller_t *entry, bool 
 
         if (!info)
             return;
-        fprintf(info->file, "# stored time: %" PRIu64 "\n", time_current(CLOCK_REALTIME));
+        fprintf(info->file, "# stored time: %" PRIu64 "\n", (uint64_t)time_current(CLOCK_REALTIME));
         // FIXME: It seems harmless, but entry->sec_keys_nw_info.pan_version and
         //        entry->sec_keys_nw_info.lpan_version are not set on wsnode.
         //        They could be replaced by ws_info.pan_information.pan_version
